Fix github comments
