+++ oak/proto/manager.proto
@@ -33,7 +33,7 @@ message ApplicationConfiguration {
   // Indication of what configuration the initial Node should run.  Must identify a
   // NodeConfiguration entry that holds a WebAssemblyConfiguration object.
   string initial_node = 2;
-  // Port number used by the gRPC pseudo-node. If not set, port will be chosen automatically.
+  // Port number used by the gRPC pseudo-node. If zero or unset, port will be chosen automatically.
   int32 grpc_port = 3;
 }
